BTT-180 Fix bug preventing deletion of theft reports without images

+++ backend/src/main/java/com/rkrs/bikethefttracker/service/ImageStorageService.java
@@ -42,6 +42,12 @@ public class ImageStorageService {
 
     public boolean deleteImages(UUID id) {
         File file = new File(PROJECT_ROOT + "/images/theft-reports/" + id);
+
+        //no files to delete
+        if (!file.exists()) {
+            return true;
+        }
+
         return FileSystemUtils.deleteRecursively(file);
     }
